fix: fix casting for handle lower/hyperhex when fetching from va_list

+++ handle_lowerhex.c
@@ -6,7 +6,7 @@
 /*   By: rfoo <rfoo@student.42singapore.sg>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/01/14 17:48:58 by rfoo              #+#    #+#             */
-/*   Updated: 2026/01/20 12:14:18 by rfoo             ###   ########.fr       */
+/*   Updated: 2026/01/20 16:24:36 by rfoo             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 int	handle_lowerhex(va_list *args)
 {
-	unsigned long	n;
+	unsigned int	n;
 
-	n = va_arg(*args, unsigned long);
-	return (print_hex(n, 0));
+	n = va_arg(*args, unsigned int);
+	return (print_hex((unsigned long)n, 0));
 }
